Compatability issues with some methods

+++ src/main/java/com/his/repository/PrescriptionRepository.java
@@ -30,17 +30,17 @@ public class PrescriptionRepository {
                 .setParameter("patientID",user.getPatient().getId())
                 .getResultList();
     }
-    public Integer getCount(User user){
-        Integer count;
+    public Long getCount(User user){
+        Long count;
         Long patientID = user.getPatient().getId();
-        return entityManager.createQuery("SELECT COUNT(p) FROM Prescription p WHERE patientID=:patientID", Integer.class)
+        return entityManager.createQuery("SELECT COUNT(p) FROM Prescription p WHERE patientID=:patientID", Long.class)
                 .setParameter("patientID",user.getPatient().getId()).getSingleResult();
     }
 
-    public Integer getOrderedCount(User user){
-        Integer count;
+    public Long getOrderedCount(User user){
+        Long count;
         Long patientID = user.getPatient().getId();
-        return entityManager.createQuery("SELECT COUNT(p) FROM Prescription p WHERE patientID=:patientID AND status='Ordered'", Integer.class)
+        return entityManager.createQuery("SELECT COUNT(p) FROM Prescription p WHERE patientID=:patientID AND status='Ordered'", Long.class)
                 .setParameter("patientID",user.getPatient().getId()).getSingleResult();
     }
 }
